requirements.txt
streamlit
pyvi
rank_bm25
google-generativeai
python-dotenv